requirements.txt
python-dateutil>=2.8.2
Pillow>=9.0.0
plyer>=2.1.0 